+++ src/main/java/ca/vanhebron/restapi/entities/Service.java
@@ -22,9 +22,8 @@ import javax.persistence.Id;
 @JsonIgnoreProperties({"hibernateLazyInitializer", "handler"})
 public class Service {
 	@Id
-	@GeneratedValue(strategy= GenerationType.AUTO)
-	private Integer id;
+	@GeneratedValue(strategy = GenerationType.AUTO)
+	private Long id;
 	private String name;
 	private String korean;
-
 }
